Hefty tests
Hefty tests are a method of defining a test that should not be run on every
patch of a series.  Testing every patch in a series can be a large resource
commitment, and in a lot of cases should just be a smoke test for
bisectability.

If "hefty = true" is set in a job, then it will only be run on "complete"
series - a single patch, or the last patch in a series and its dependencies.

Signed-off-by: Russell Currey <ruscur@russell.cc>
[ajd: rebase on serde changes]
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>